+++ C-Project/src/main.c
@@ -29,7 +29,7 @@ int main(const int argc, const char* argv[]) {
 
     FILE* ppm;
     uint8_t* ppm_data;
-    const unsigned int ppm_data_size = ppm_init(out_path, image_width, image_height, &ppm, &ppm_data);
+    const size_t ppm_data_size = ppm_init(out_path, image_width, image_height, &ppm, &ppm_data);
 
     scene_destroy(scene);
     ppm_end(ppm, ppm_data, ppm_data_size);
